+++ src/commands.js
@@ -1,5 +1,6 @@
 import { up, cd, ls, cat, add, rn } from './filesystem.js';
 import { printInfo } from './os-info.js';
+import { printOperationFailed } from './console.js';
 
 const Commands = {
     UP: 'up',
@@ -8,6 +9,7 @@ const Commands = {
     CAT: 'cat',
     ADD: 'add',
     RN: 'rn',
+    CP: 'cp',
 
     OS: 'os',
 };
@@ -33,7 +35,7 @@ export async function handleCommand(str) {
             await add(args);
             break;
         case Commands.RN:
-            await rn(args);
+            await rn(extractTwoArgs(args));
             break;
         case Commands.OS:
             printInfo(args);
@@ -42,8 +44,7 @@ export async function handleCommand(str) {
 }
 
 /**
- *
- * @param str
+ * @param str {string}
  * @returns {{args: string, command: string}}
  */
 function extractCommand(str) {
@@ -54,5 +55,22 @@ function extractCommand(str) {
     return {
         command,
         args,
+    };
+}
+
+/**
+ * @param args {string}
+ * @returns {{first: string, second: string}}
+ */
+function extractTwoArgs(args) {
+    const arr = args.split(' ');
+
+    if (arr.length < 2) {
+        printOperationFailed();
+    }
+
+    return {
+        first: arr[0].trim(),
+        second: arr[1].trim(),
     };
 }
